Filter the Table Data
Snowflake Badge 2: Data Application Builders Workshop
Lesson 3: A Quick and Easy Streamlit App!  🥋 Connecting Our Table to the Pick List

We'll ask our app to put the list of selected fruits into a variable called fruits_selected. Then, we'll ask our app to use the fruits in our fruits_selected list to pull rows from the full data set (and assign that data to a variable called fruits_to_show). Finally, we'll ask the app to use the data in fruits_to_show in the dataframe it displays on the page.